-change: introduce an alternative implementation of async sh2 processing which does a real perf improvement only syncing when necessary. (WIP and has rare glitches but runs a bit too fast for now)

+++ yabause/src/ctrl/include/yabause.h
@@ -148,7 +148,7 @@ extern void resetFrameSkip();
 
 extern u32 saved_scsp_cycles;
 #define SCSP_FRACTIONAL_BITS 20
-u32 get_cycles_per_line_division(u32 clock, int frames, int lines, int divisions_per_line);
+u64 get_cycles_per_line_division(u32 clock, int frames, int lines, int divisions_per_line);
 u32 YabauseGetCpuTime();
 
 #ifdef __cplusplus
